backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
kiteconnect==5.0.1
asyncpg==0.30.0
httpx==0.27.0
anthropic==0.40.0
numpy==2.1.0
scipy==1.14.0
python-dotenv==1.0.1
